Validation added

+++ BikeRepository.cs
@@ -19,35 +19,45 @@ namespace BikeRentalManagement_V_2
         {
             Console.Write("Enter ID:");
             string Id = Console.ReadLine();
-
-            Console.Write("Enter Brand:");
-            string brand = Console.ReadLine();
-            string cap=Capitalize(brand);
-
-            Console.Write("Enter Model:");
-            string model = Console.ReadLine();
-
-            Console.Write("Enter Renal:");
-            decimal price = decimal.Parse(Console.ReadLine());
-
-            Bike bike = new Bike(Id, brand, model, price);
-
-            var query = @"Insert into BikesData (BikeId,Model,Brand,RentalPrice) values(@id,@brand,@model,@price)";
-
-            using (var connection = new SqlConnection(connectionstring))
+            bool chkId = checkID(Id);
+            if(!chkId)
             {
-                using (var command = new SqlCommand(query, connection))
-                {
-                    command.Parameters.AddWithValue("@id", Id);
-                    command.Parameters.AddWithValue("@brand", cap);
-                    command.Parameters.AddWithValue("@model", model);
-                    command.Parameters.AddWithValue("@price", price);
+                Console.Write("Enter Brand:");
+                string brand = Console.ReadLine();
+                string cap = Capitalize(brand);
 
-                    connection.Open();
-                    command.ExecuteNonQuery();
-                    Console.WriteLine("Added");
+                Console.Write("Enter Model:");
+                string model = Console.ReadLine();
+
+                Console.Write("Enter Renal:");
+                decimal price = decimal.Parse(Console.ReadLine());
+
+                Bike bike = new Bike(Id, brand, model, price);
+
+                var query = @"Insert into BikesData (BikeId,Model,Brand,RentalPrice) values(@id,@brand,@model,@price)";
+
+                using (var connection = new SqlConnection(connectionstring))
+                {
+                    using (var command = new SqlCommand(query, connection))
+                    {
+                        command.Parameters.AddWithValue("@id", Id);
+                        command.Parameters.AddWithValue("@brand", cap);
+                        command.Parameters.AddWithValue("@model", model);
+                        command.Parameters.AddWithValue("@price", price);
+
+                        connection.Open();
+                        command.ExecuteNonQuery();
+                        Console.WriteLine("Added");
+                    }
                 }
+
             }
+            else
+            {
+                Console.WriteLine("Already Id exist")
+            }
+
+          
 
 
         }
@@ -75,13 +85,36 @@ namespace BikeRentalManagement_V_2
         {
             Console.Write("Enter Id:");
             string Id=Console.ReadLine();
-            var query = @"Delete from BikesData where BikeId=@Id";
-            using(var connection = new SqlConnection(connectionstring))
+            bool chkId = checkID(Id);
+            if (chkId)
             {
-                using(var command = new SqlCommand(query,connection))
+                var query = @"Delete from BikesData where BikeId=@Id";
+                using (var connection = new SqlConnection(connectionstring))
                 {
-                    command.Parameters.AddWithValue("@Id", Id);
-                    command.ExecuteNonQuery ();
+                    using (var command = new SqlCommand(query, connection))
+                    {
+                        command.Parameters.AddWithValue("@Id", Id);
+                        command.ExecuteNonQuery();
+                    }
+                }
+
+            }
+            else
+            {
+                Console.Write("Invalid");
+            }
+         
+        }
+        public bool checkID(string id)
+        {
+            var query = @"select count(1) from BikesData where BikeId=@Id";
+            using(var connection=new SqlConnection(connectionstring))
+            {
+                using(var command=new SqlCommand(query,connection))
+                {
+                   command.Parameters.AddWithValue("Id",id);
+                    int result =(int)command.ExecuteScalar();
+                    return result > 0;
                 }
             }
         }
@@ -90,31 +123,40 @@ namespace BikeRentalManagement_V_2
         {
             Console.Write("Enter Id:");
             string Id=Console.ReadLine();
-
-            Console.Write("Enter Brand:");
-            string brand = Console.ReadLine();
-
-
-            Console.Write("Enter Model:");
-            string model = Console.ReadLine();
-
-            Console.Write("Enter Renal:");
-            decimal price = decimal.Parse(Console.ReadLine());
-
-            Bike bike = new Bike(Id, brand, model, price);
-
-            var query = @"update BikesData set Brand=@brand,Model=@model,RentalPrice=@price where BikeId=@Id";
-            using (var connection = new SqlConnection(connectionstring))
+            bool chkId=checkID(Id);
+            if (chkId)
             {
-                using (var command = new SqlCommand(query, connection))
+                Console.Write("Enter Brand:");
+                string brand = Console.ReadLine();
+
+
+                Console.Write("Enter Model:");
+                string model = Console.ReadLine();
+
+                Console.Write("Enter Renal:");
+                decimal price = decimal.Parse(Console.ReadLine());
+
+                Bike bike = new Bike(Id, brand, model, price);
+
+                var query = @"update BikesData set Brand=@brand,Model=@model,RentalPrice=@price where BikeId=@Id";
+                using (var connection = new SqlConnection(connectionstring))
                 {
-                    command.Parameters.AddWithValue("@Id", Id);
-                    command.Parameters.AddWithValue("@brand", brand);
-                    command.Parameters.AddWithValue("@model", model);
-                    command.Parameters.AddWithValue("@price", price);
-                    command.ExecuteNonQuery();
+                    using (var command = new SqlCommand(query, connection))
+                    {
+                        command.Parameters.AddWithValue("@Id", Id);
+                        command.Parameters.AddWithValue("@brand", brand);
+                        command.Parameters.AddWithValue("@model", model);
+                        command.Parameters.AddWithValue("@price", price);
+                        command.ExecuteNonQuery();
+                    }
                 }
             }
+            else
+            {
+                Console.Write("Invalid ID");
+            }
+
+            
         }
     }
 }
